Rename typo in filename
