+++ src/client/components/highlight-modal/highlight-modal.tsx
